requirements-dev.txt
dj-static==0.0.6
Django==2.1.1
django-enumfields==0.10.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
static3==0.7.0
